+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Docker action that didn't push 2 tags for a new release, just the "latest"
 - Dockerfile faster to build and cleaner code
 - Modified jwt.decode params to be compliant to PyJWT v2.0
+- Bug when trying to delete variants for samples not in dataset
 
 
 ## [1.4] - 2020.11.05